#14 move headers with 'public' functions in include directory
