+++ src/ev3_output.c
@@ -15,7 +15,7 @@
  *
  */
 
-#include "ev3_output.h"
+#include "../include/ev3_output.h"
 
 #define MIN(a, b) (((a)<(b))?(a):(b))
 #define MAX(a, b) (((a)>(b))?(a):(b))
