+++ src/ev3_sensors/ev3_gyro.c
@@ -1,6 +1,6 @@
 #include "../ev3_inputs/ev3_input_uart.h"
 #include "../../firmware_headers/lms2012/ev3_basictypes.h"
-#include "../ev3_wait.h"
+#include "../../include/ev3_wait.h"
 #include "ev3_gyro.h"
 
 #define EV3_GYRO_SENSOR_TYPE                32
