+++ src/ev3_sensors/ev3_ir.c
@@ -1,5 +1,5 @@
 #include "../ev3_inputs/ev3_input_uart.h"
-#include "../ev3_wait.h"
+#include "../../include/ev3_wait.h"
 #include "ev3_ir.h"
