+++ src/ev3_sensors/ev3_sensors.c
@@ -1,5 +1,5 @@
 #include <stddef.h>
-#include "../ev3_constants.h"
+#include "../../include/ev3_constants.h"
 #include "ev3_sensors.h"
