+++ src/ev3_sensors/ev3_touch.h
@@ -1,20 +1,13 @@
-#ifndef EV3_API_EV3_TOUCH_SENSOR_H
-#define EV3_API_EV3_TOUCH_SENSOR_H
+#ifndef EV3_API_EV3_TOUCH_SENSOR_PRIVATE_H
+#define EV3_API_EV3_TOUCH_SENSOR_PRIVATE_H
 
 #include <stdbool.h>
-#include "ev3_sensors.h"
-
-#define EV3_TOUCH_SENSOR_TOUCH_MODE 0
-
-extern SensorHandler * EV3Touch;
+#include "../../include/ev3_sensors/ev3_touch.h"
 
 bool initEV3Touch(int port);
 
-bool ReadEV3TouchSensor(int sensorPort);
-
 void exitEV3Touch(int port);
 
-/* Utility functions */
 void setEV3UltrasonicSensorMode(int port, int mode);
 
-#endif //EV3_API_EV3_TOUCH_SENSOR_H
+#endif
