+++ src/ev3_sensors/ht_compass.h
@@ -1,24 +1,9 @@
-#ifndef EV3_API_HT_COMPASS_H
-#define EV3_API_HT_COMPASS_H
+#ifndef EV3_API_HT_COMPASS_PRIVATE_H
+#define EV3_API_HT_COMPASS_PRIVATE_H
 
-#include "ev3_sensors.h"
-
-#define HT_COMPASS_SENSOR_DEFAULT_MODE 0
-
-extern SensorHandler * HTCompass;
-
-typedef enum HTCompassReadingMode {
-    HTCompassCompass,
-    HTCompassAngle
-} HTCompassReadingMode;
+#include "../../include/ev3_sensors/ht_compass.h"
 
 bool initHTCompassSensor(int port);
-
-int ReadHTCompassSensor(int port, HTCompassReadingMode mode);
-
-void StartHTCompassCalibration(int port);
-
-void StopHTCompassCalibration(int port);
 
 void exitHTCompassSensor(int port);
