+++ src/ev3_sensors/ht_ir.h
@@ -1,25 +1,13 @@
-#ifndef EV3_API_HT_IR_H
-#define EV3_API_HT_IR_H
+#ifndef EV3_API_HT_IR_PRIVATE_H
+#define EV3_API_HT_IR_PRIVATE_H
 
-#include "ev3_sensors.h"
+#include "../../include/ev3_sensors/ev3_ir.h"
 
-#define HT_IR_SENSOR_DEFAULT_MODE 0
-
-// TODO: Specify version 2
-extern SensorHandler * HTIr;
-
-typedef enum HTIrReadingMode {
-    Modulated,
-    Unmodulated
-} HTIrReadingMode;
 
 bool initHTIrSensor(int port);
 
-int ReadHTIrSensor(int port, HTIrReadingMode mode);
-
 void exitHTIrSensor(int port);
 
-/* Utility functions */
 int getHTIrRegisterForReadingMode(HTIrReadingMode mode);
 
-#endif //EV3_API_HT_IR_H
+#endif
