+++ src/ev3_sensors/pixy_cam.h
@@ -1,21 +1,13 @@
-#ifndef EV3_API_PIXY_CAM_H
-#define EV3_API_PIXY_CAM_H
+#ifndef EV3_API_PIXY_CAM_PRIVATE_H
+#define EV3_API_PIXY_CAM_PRIVATE_H
 
 #include <stdbool.h>
 #include "ev3_sensors.h"
 
 extern SensorHandler * PixyCam;
 
-typedef struct PixyRectangle {
-    int signature;
-    int x, y;
-    int width, height;
-    int angle;
-} PixyRectangle;
 
 bool initPixyCamSensor(int port);
-
-PixyRectangle ReadPixyCamSensorBiggestBlobRectangle(int port);
 
 void exitPixyCamSensor(int port);
