+++ src/ev3_timer.c
@@ -15,7 +15,7 @@
  *
  */
 
-#include "ev3_timer.h"
+#include "../include/ev3_timer.h"
 
 #ifndef DISABLE_TIMERS
