+++ src/ev3_wait.c
@@ -15,19 +15,17 @@
  *
  */
 
-#include "ev3_wait.h"
+#include "../include/ev3_wait.h"
 #include <time.h>
 
-static void ms2ts(struct timespec *ts, unsigned long ms)
-{
-	ts->tv_sec = ms / 1000;
-	ts->tv_nsec = (ms % 1000) * 1000000;
+static void ms2ts(struct timespec *ts, unsigned long ms) {
+    ts->tv_sec = ms / 1000;
+    ts->tv_nsec = (ms % 1000) * 1000000;
 }
 
-void Wait(unsigned long ms)
-{
-	struct timespec delay;
-	ms2ts(&delay, ms);
-	while (nanosleep(&delay, &delay));
+void Wait(unsigned long ms) {
+    struct timespec delay;
+    ms2ts(&delay, ms);
+    while (nanosleep(&delay, &delay));
 }
